==> in-silico-approach-on-herbal-compound-identification/src/2-principal-component-analysis.ipynb <==
import pandas as pd
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

# 1 Import Data

dataset = pd.read_csv("../data/results/all_fingerprints.csv")
dataset.head()

# 2. Explore Data

dataset.info()

dataset.describe()

dataset[dataset["class"] == 0].info() # 2322
print()
dataset[dataset["class"] == 1].info() # 78
print()
dataset[dataset["class"] == 2].info() # 5396

# 3 Data Cleansing

dataset.drop(['smiles'], axis=1, inplace=True)
dataset.head(10)

dataset.tail(10)

# 4 Data Splitting

X = dataset.drop("class", axis=1)
y = dataset["class"]

X

y

# 5 Standardized Features

scaler = StandardScaler()
X_std = scaler.fit_transform(X)

# 6 PCA

pca = PCA(n_components=0.9) # Keep 90% of Informations
X_pca = pca.fit_transform(X_std)

X_pca

X_pca.shape

X_pca_df = pd.DataFrame(X_pca)

X_pca_df

# 7 Concatenate with Target

X_pca_df["class"] = y
pca_df = X_pca_df
pca_df.head(10)

pca_df[pca_df["class"] == 0].info() # 2322
print()
pca_df[pca_df["class"] == 1].info() # 78
print()
pca_df[pca_df["class"] == 2].info() # 5396

# 8 Save to CSV

pca_df.to_csv("../data/pca/pca_fingerprint_results.csv", index=False)
